# desktop/openapi.json
{
  "components": {
    "schemas": {
...
      "Assumption": {
        "description": "可解释推断；Alpha 优先 user_authorized / planseed_default。",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "source": {
            "default": "llm_inference",
            "enum": [
              "user_authorized",
              "planseed_default",
              "llm_inference"
            ],
            "title": "Source",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "Assumption",
        "type": "object"
      },
...
      "DesignPreferences": {
        "properties": {
          "prefer_compact_footprint": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefer Compact Footprint"
          },
          "prefer_open_kitchen_dining": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefer Open Kitchen Dining"
          },
          "prefer_short_corridor": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefer Short Corridor"
          },
          "prefer_south_facing_living": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefer South Facing Living"
          },
          "quiet_zone_away_from_entry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Zone Away From Entry"
          },
          "wet_stack_preference": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wet Stack Preference"
          }
        },
        "title": "DesignPreferences",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HouseholdRequirements": {
        "properties": {
          "bathrooms": {
            "anyOf": [
              {
                "maximum": 8.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bathrooms"
          },
          "bedrooms": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms"
          },
          "has_garage": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Garage"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "occupants": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupants"
          }
        },
        "title": "HouseholdRequirements",
        "type": "object"
      },
...
      "ParseNLRequest": {
        "properties": {
          "max_repairs": {
            "default": 2,
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Max Repairs",
            "type": "integer"
          },
          "text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ParseNLRequest",
        "type": "object"
      },
      "ParseNLResponse": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "raw": {
            "additionalProperties": true,
            "title": "Raw",
            "type": "object"
          },
          "repair_notes": {
            "items": {
              "type": "string"
            },
            "title": "Repair Notes",
            "type": "array"
          },
          "requirement_spec": {
            "$ref": "#/components/schemas/RequirementSpec"
          }
        },
        "required": [
          "requirement_spec",
          "attempts",
          "provider"
        ],
        "title": "ParseNLResponse",
        "type": "object"
      },
...
      "ProjectDetail": {
        "properties": {
          "current_evaluation_version": {
            "title": "Current Evaluation Version",
            "type": "string"
          },
          "evaluation_version_mismatch": {
            "default": false,
            "title": "Evaluation Version Mismatch",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ProjectPayload"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "updated_at",
          "payload"
        ],
        "title": "ProjectDetail",
        "type": "object"
      },
      "ProjectMeta": {
        "properties": {
          "app_version": {
            "default": "0.1.0",
            "title": "App Version",
            "type": "string"
          },
          "format_version": {
            "default": "1",
            "title": "Format Version",
            "type": "string"
          }
        },
        "title": "ProjectMeta",
        "type": "object"
      },
...
      "ProjectPayload": {
        "properties": {
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Candidates",
            "type": "array"
          },
          "compare_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare Id"
          },
          "form": {
            "additionalProperties": true,
            "title": "Form",
            "type": "object"
          },
          "locks": {
            "additionalProperties": true,
            "title": "Locks",
            "type": "object"
          },
          "program": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          },
          "project_meta": {
            "$ref": "#/components/schemas/ProjectMeta"
          },
          "requirement_spec": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Phase 5.1.1：canonical RequirementSpec；不可用 ProgramSummary 替代",
            "title": "Requirement Spec"
          },
          "schema_versions": {
            "$ref": "#/components/schemas/SchemaVersions"
          },
          "selected_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Id"
          }
        },
        "title": "ProjectPayload",
        "type": "object"
      },
      "ProjectSummaryOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "updated_at"
        ],
        "title": "ProjectSummaryOut",
        "type": "object"
      },
...
      "RequirementSpec": {
        "description": "用户自然语言或表单表达的需求。\n\nRequirementSpec → normalize → DesignProgram",
        "properties": {
          "assumptions": {
            "items": {
              "$ref": "#/components/schemas/Assumption"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "floor_count": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Count"
          },
          "household": {
            "$ref": "#/components/schemas/HouseholdRequirements"
          },
          "preferences": {
            "$ref": "#/components/schemas/DesignPreferences"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "relation_intents": {
            "description": "Phase 6：名称级邻接/通行意图；Normalizer 可暂忽略",
            "items": {
              "$ref": "#/components/schemas/RelationIntent"
            },
            "title": "Relation Intents",
            "type": "array"
          },
          "site": {
            "$ref": "#/components/schemas/SiteRequirements"
          },
          "spaces": {
            "items": {
              "$ref": "#/components/schemas/SpaceRequirement"
            },
            "title": "Spaces",
            "type": "array"
          },
          "unknowns": {
            "items": {
              "$ref": "#/components/schemas/UnknownRequirement"
            },
            "title": "Unknowns",
            "type": "array"
          }
        },
        "title": "RequirementSpec",
        "type": "object"
      },
...
      "SaveProjectRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ProjectPayload"
          }
        },
        "required": [
          "name",
          "payload"
        ],
        "title": "SaveProjectRequest",
        "type": "object"
      },
      "SchemaVersions": {
        "description": "快照内设计溯源；Save 不得仅因保存而改写为 current。",
        "properties": {
          "assignment_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Strategy"
          },
          "evaluation_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Version"
          },
          "generator_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generator Strategy"
          },
          "generator_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generator Version"
          },
          "geometry_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry Backend"
          },
          "selection_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Strategy"
          },
          "selection_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Version"
          },
          "solver_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solver Version"
          }
        },
        "title": "SchemaVersions",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PlanSeed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummaryOut"
                  },
                  "title": "Response List Projects Api Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "save_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/import": {
      "post": {
        "description": "打开 / 导入 `.planseed`：请求体为 ZIP 字节。默认拒绝覆盖已有 id。",
        "operationId": "import_planseed_package_api_projects_import_post",
        "parameters": [
          {
            "in": "query",
            "name": "overwrite",
            "required": false,
            "schema": {
              "default": false,
              "title": "Overwrite",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Planseed Package",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_api_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Project Api Projects  Project Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{project_id}/package": {
      "get": {
        "description": "导出已保存项目为 `.planseed` ZIP。",
        "operationId": "export_planseed_package_api_projects__project_id__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Planseed Package",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/requirements/parse": {
      "post": {
        "description": "自然语言 → RequirementSpec（含有限 repair；无几何）。",
        "operationId": "parse_requirements_nl_api_requirements_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseNLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseNLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Requirements Nl",
        "tags": [
          "requirements"
        ]
      }
    }
  }
}